Created skeleton for magazine subsystem, added Utilities.java
The magazine subsystem as of now should be able to run the actual mechanism. There aren't controls yet, though. The Utilities.java file is for useful functions that we'll use throughout the code. There's only one in there right now, for clipping values between a minimum and maximum value, but as we need to use other functions (like finding the distance between two angles, interpolating between values, etc.), we can add them in.