TurnCommand works
I'm not sure what wasn't working before, but I just got rid of a lot of variables that weren't necessary, used doubles instead of Rotation2d objects, and replaced the piecewise function with a single function that seems to work pretty well for large and small angles. It can't do the thing where it turns more than 180 degrees if you want it to, but I really don't think it's that necessary (I mean, for the navigation-based challenges you could even turn so that the robot finishes the turn facing backwards, since it drives as well backwards as it does forwards and the robot would never need to turn more than 90 degrees, but I feel like that might overcomplicate things).