Possibly fixed unit conversion error with maximum hood position, fixed controls to measure in radians
I made more of the variables about the hood's range public, so commands can use them to make sure that the requests they're making actually make sense. Also, I'm pretty sure that the value I found for the maximum hood position was in radians and not rotations, which might explain why the hood tried to move past its maximum on Saturday. I got rid of the conversion factor from rotations to radians, so hopefully that will fix it.