Flywheel works with power, PID fixed for the hood
There were a bunch of dumb problems with the shooter, and I think they're all fixed now? The hood can rotate to angles, and the flywheel can run (only in PercentOutput mode so far, but I do have a value of kF now). We tested the flywheel at 75% power, and it was more than able to slam power cells into the ceiling. I guess what's next is to get PID working for the flywheel and the turret, and then we're good?