Miscellaneous control changes to make flywheel testing easier
I switched the buttons around so that none of the controls overlap, and you could in theory run every every mechanism on the robot manually with the operator controls. Considering that you'd need to be using both triggers, the X,Y,A,and (optionally) B buttons, the left bumper, and the start and back buttons to do so, that's _probably_ not a great idea (at least until it's more automated) but at least none of the controls are too weird or unintuitive.