TurretPortAlignCommand and TurretPositionCommand both work, added slew rate limiter to the turret
The turret can now move to different positions (as long as it resets when it's done), and align to face the power port. The auto-aligning only works if the port tracker has data, which isn't great (probably similar to the scanning mode for the auto-chasing command for the drivetrain), but it's certainly progress.