Auto-aiming turret while shooting, made simple auto-firing command using turret, shooter, and magazine
The code just runs the turret auto-aligning while in parallel setting the flywheel and hood angle, advancing the magazine by four power cell diameters, and then turning the flywheel off and setting the hood back down. It's really simplistic (ideally the robot should determine on its own when it's done firing based on the flywheel speed or something, and it has no way of knowing what hood angle and flywheel velocity to set), but it does work, and that's kind of exciting.