Removed unnecessary data point from InterpolatorTables at 3.02m, added some simple edits to improve debuggability
The data point for 3.02 meters was not necessary (there's already one at 3.05 meters), and looking at the data it's clearly anomalous (its hood angle is the same as the next and previous entries, but the flywheel speed is actually higher than the flywheel speed for 3.05m despite being closer). I got rid of that, and I also added in some extra printouts to confirm that TargetShooterCommand and TargetHoodCommand are both working properly.