Made it so that the flywheel no longer runs continuously
I made a (possibly temporary) change to the TargetFlywheelCommand so that it doesn't get confused by power cells passing in front of the range sensor. Also, it prevents significant disagreement between the range the hood and the flywheel are using, which might be problematic for driving on the move. Ideally, both commands should run continuously, but that would represent a significant restructuring of the code. I might create a new branch for that to see if I can make it work (I would probably put the magazine and power down commands in their own branch of the command group, make it so the hood never finishes aligning, and make a third wait command (WaitToAdvance) that only finishes when the hood, flywheel, and port tracker are all aligned right before running the magazine. I'd also need to add in detectors to ignore abrupt jumps in the range values for the hood and flywheel. That way, the robot would be able to adjust everything right up until firing, and the robot would just 'blink' when the power cell gets in the way). Okay, almost definitely going to do that (as long as people don't think it's a bad idea).