Fixed some things
The projectedDistance calculation was using min instead of max, so it would've been only able to drive backwards. Also, I got rid of the clamping on the rotation curve. That might not be a good idea, since it means that the rotation speed scale factor isn't the maximum speed anymore, but it will let the robot turn faster if it's really far off, and it still won't go dangerously fast or anything.