Galactic Search auto almost works, collector is probably haunted
We made a ton of changes to get it working consistently, but the collector has some serious problems. We got something submittable, but hopefully once we have more time to test it and improve anything left.

Important note: the magazine controls _do_ let you move it downwards, but after we damaged the RSL wire doing that by mistake I decided to cut down the distance per press by a factor of 10, so we can't break anything in that particular way.