Shooter is (sort of) able to fire a power cell without firing the next one
I improved the PID tuning for the magazine and switched to using the hood position and some linear filters to identify when the power cell is fired. It still be improved a lot, though, but it can fire without awkwardly sort of firing the second one (or awkwardly sort of firing both).