Added one-button firing and fixed some errors with WaitForShooterCurrentSpike and ShooterSetCommand
It's ever-so-slightly janky, but it actually works. Also, I just scaled the flywheel velocities in the InterpolatorTable for the lower port down by the factor the high port is from the low port (since the low port table works for the high port). It actually works, and it shoots into the port with a really similar profile to how it works for the higher port, so  that's promising.