Increased flywheel velocity
I was going to uncomment the code for recentering the flywheel, but I already did that (just not with the code we had deployed). I'm committing now just because I have a non-trivially janky solution to the issue where the autofiring routine can't be interrupted, and I don't want this change to get caught up with that.