Added a somewhat more functional version of the left 5-cell auto
The issue now is that it can't fire that far away from its endpoint (and it can't even see the power port from there so there isn't much hope of firing from there), and if we drive backwards we'd run out of time. I guess we either settle for a 4-cell auto or make some part of it faster (the two things that we could do now would be to increase the  rate limiter on the flywheel and deploy the collector in parallel with firing the first three power cells, but there might be more?).