Got rid of redundant definitions
I don't know if it's less confusing to have the variables defined at the top and not used until later or to be defined far away from their displays, but I'm guessing it's the second one? Also, it's really confusing having both of them, so getting rid of one definition seemed prudent. I was considering adding in some other stuff to the demo (input buttons, stuff like that), but explaining those requires explaining a lot of Java stuff that might make it more intimidating for people who haven't learned it yet (or at the very least, less helpful).